# AI Generated Gnome API - OpenAPI 3.0.yaml
openapi: 3.0.3
info:
  title: Gnome Board Game Shop API
  description: |
    RESTful API for managing board game products, categories, variants, and admin authentication.
    
    ## Authentication
    This API uses JWT Bearer token authentication for admin operations. 
    - Public endpoints (GET) don't require authentication
    - Admin endpoints (POST, PUT, DELETE) require valid JWT token with Admin role
    
    ## Getting Started
    1. Use the `/auth/login` endpoint to authenticate and get JWT token
    2. Include the JWT token in the Authorization header for protected endpoints
    3. Use `/auth/refresh` to get new tokens when the current one expires
    
    ## Rate Limiting
    - 100 requests per minute per IP address
    - 1000 requests per hour per authenticated user
    
    ## Error Handling
    All errors return a consistent format with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: Gnome Board Game Shop
    email: support@gnomeboardgames.com
    url: https://gnomeboardgames.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://localhost:7080
    description: Development server
  - url: https://api.gnomeboardgames.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /product/list:
    get:
      tags:
        - Products
      summary: Get products list
      description: Retrieves a paginated list of products with optional filtering by name, date range, and sorting options
      operationId: listProducts
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page-size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: date-from
          in: query
          description: Filter products created from this date (YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
        - name: date-to
          in: query
          description: Filter products created until this date (YYYY-MM-DD)
          required: false
          schema:
            type: string
            format: date
        - name: name
          in: query
          description: Filter products by name (partial match)
          required: false
          schema:
            type: string
        - name: sort-by
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum: [created-date-time, name, updated-date-time]
            default: created-date-time
        - name: sort-order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
              examples:
                products_list:
                  summary: Products list response
                  value:
                    items:
                      - id: 1
                        name: "Chess Master Pro"
                        slug: "chess-master-pro"
                        description: "Professional chess set with premium pieces"
                        categories:
                          - id: 1
                            name: "Strategy Games"
                            slug: "strategy-games"
                        variants:
                          - id: 1
                            name: "Standard Edition"
                            price: 29.99
                            stock: 50
                            isPrimary: true
                        createdDateTime: "2024-01-01T00:00:00Z"
                        updatedDateTime: "2024-01-15T10:00:00Z"
                    totalCount: 1
                    pageNumber: 1
                    pageSize: 10
                    totalPages: 1
                    hasPreviousPage: false
                    hasNextPage: false
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /product/{slug}:
    get:
      tags:
        - Products
      summary: Get product by slug
      description: Retrieves detailed information about a specific product including its categories and variants using its slug
      operationId: getProductBySlug
      parameters:
        - name: slug
          in: path
          description: The unique slug identifier of the product
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9-]+$'
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
              examples:
                product_details:
                  summary: Product details response
                  value:
                    id: 1
                    name: "Chess Master Pro"
                    slug: "chess-master-pro"
                    description: "Professional chess set with premium pieces"
                    categories:
                      - id: 1
                        name: "Strategy Games"
                        slug: "strategy-games"
                    variants:
                      - id: 1
                        name: "Standard Edition"
                        slug: "standard-edition"
                        price: 29.99
                        stock: 50
                        isPrimary: true
                        imageUrl: "https://res.cloudinary.com/example/image/upload/v1/chess-standard.jpg"
                    createdDateTime: "2024-01-01T00:00:00Z"
                    updatedDateTime: "2024-01-15T10:00:00Z"
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Products
      summary: Delete product
      description: Permanently deletes a board game product and all its variants. Requires admin authentication.
      operationId: deleteProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: The unique identifier of the product to delete
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDeletedResponse'
              examples:
                deletion_success:
                  summary: Successful deletion
                  value:
                    success: true
        '401':
          description: Unauthorized - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /product/add:
    post:
      tags:
        - Products
      summary: Add new product
      description: Creates a new board game product with specified details and category associations. Requires admin authentication.
      operationId: addProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AddProductRequest'
            examples:
              new_product:
                summary: New product creation
                value:
                  name: "New Board Game"
                  slug: "new-board-game"
                  description: "An exciting new board game"
                  categoryIds: "1,2"
      responses:
        '200':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCreatedResponse'
              examples:
                creation_success:
                  summary: Successful creation
                  value:
                    id: 2
        '400':
          description: Invalid product data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /product/update:
    put:
      tags:
        - Products
      summary: Update product
      description: Updates an existing board game product with new details. Requires admin authentication.
      operationId: updateProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
            examples:
              product_update:
                summary: Product update
                value:
                  id: 1
                  name: "Updated Chess Master Pro"
                  slug: "updated-chess-master-pro"
                  description: "Updated description for the chess set"
                  categoryId: 1
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductUpdatedResponse'
              examples:
                update_success:
                  summary: Successful update
                  value:
                    id: 1
        '400':
          description: Invalid product data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme. Enter your token in the text input below.
  schemas:
...
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        totalCount:
          type: integer
          description: Total number of products
          example: 1
        pageNumber:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Number of items per page
          example: 10
        totalPages:
          type: integer
          description: Total number of pages
          example: 1
        hasPreviousPage:
          type: boolean
          description: Whether there is a previous page
          example: false
        hasNextPage:
          type: boolean
          description: Whether there is a next page
          example: false
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          description: Product ID
          example: 1
        name:
          type: string
          description: Product name
          example: "Chess Master Pro"
        slug:
          type: string
          description: URL-friendly product name
          example: "chess-master-pro"
        description:
          type: string
          description: Product description
          example: "Professional chess set with premium pieces"
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryDto'
        variants:
          type: array
          items:
            $ref: '#/components/schemas/VariantDto'
        createdDateTime:
          type: string
          format: date-time
          description: Product creation date
          example: "2024-01-01T00:00:00Z"
        updatedDateTime:
          type: string
          format: date-time
          description: Product last update date
          example: "2024-01-15T10:00:00Z"
    AddProductRequest:
      type: object
      required:
        - name
        - slug
        - description
        - categoryIds
      properties:
        name:
          type: string
          description: Product name
          example: "New Board Game"
        slug:
          type: string
          description: URL-friendly product name
          example: "new-board-game"
        description:
          type: string
          description: Product description
          example: "An exciting new board game"
        categoryIds:
          type: string
          description: Comma-separated list of category IDs
          example: "1,2"
    UpdateProductRequest:
      type: object
      required:
        - id
        - name
        - slug
        - description
        - categoryId
      properties:
        id:
          type: integer
          description: Product ID to update
          example: 1
        name:
          type: string
          description: Updated product name
          example: "Updated Chess Master Pro"
        slug:
          type: string
          description: Updated URL-friendly product name
          example: "updated-chess-master-pro"
        description:
          type: string
          description: Updated product description
          example: "Updated description for the chess set"
        categoryId:
          type: integer
          description: Category ID
          example: 1
    ProductCreatedResponse:
      type: object
      properties:
        id:
          type: integer
          description: The ID of the newly created product
          example: 2
    ProductUpdatedResponse:
      type: object
      properties:
        id:
          type: integer
          description: The ID of the updated product
          example: 1
    ProductDeletedResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Confirmation that the product was deleted successfully
          example: true

    # Category Schemas
...
    CategoryDto:
      type: object
      properties:
        id:
          type: integer
          description: Category ID
          example: 1
        name:
          type: string
          description: Category name
          example: "Strategy Games"
        slug:
          type: string
          description: URL-friendly category name
          example: "strategy-games"
...
    VariantDto:
      type: object
      properties:
        id:
          type: integer
          description: Variant ID
          example: 1
        name:
          type: string
          description: Variant name
          example: "Standard Edition"
        slug:
          type: string
          description: URL-friendly variant name
          example: "standard-edition"
        price:
          type: number
          format: decimal
          description: Variant price
          example: 29.99
        stock:
          type: integer
          description: Available stock quantity
          example: 50
        isPrimary:
          type: boolean
          description: Whether this is the primary variant
          example: true
        imageUrl:
          type: string
          format: uri
          description: Variant image URL
          example: "https://res.cloudinary.com/example/image/upload/v1/chess-standard.jpg"
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message describing what went wrong
          example: "Invalid username or password" 